+++ djpcms/core/exceptions.py
@@ -1,12 +1,20 @@
 from django.core.exceptions import *
 
+
+class DjpcmsException(Exception):
+    '''Base class for ``djpcms`` related exceptions.'''
+    pass
+
 class UsernameAlreadyAvailable(Exception):
     pass
 
 
-class ApplicationUrlException(Exception):
+class ApplicationUrlException(DjpcmsException):
+    '''A :class:`DjpcmsException` raised when there are problems
+related to urls configuration'''
     pass
 
 
-class PageNotFound(Exception):
+class PageNotFound(DjpcmsException):
+    '''A :class:`DjpcmsException` raised when page is not found.'''
     pass