+++ djpcms/plugins/info.py
@@ -22,13 +22,14 @@ class PoweredBy(DJPplugin):
                                         c)
         
 class Deploy(DJPplugin):
+    '''Information about deployment. Based on :class:`djpcms.contrib.jdep.models.DeploySite`.
+If `djpcms.contrib.jdep` is not part of your ``INSTALLED_APPS``
+nothing will be displayed.
+    '''
     name = 'deploysite'
     description = 'Deploy Timestamp'
     
     def render(self, djp, wrapper, prefix, **kwargs):
-        '''
-        Information about deployment
-        '''
         from djpcms.conf import settings
         from djpcms.utils import timezone
         if 'djpcms.contrib.jdep' in settings.INSTALLED_APPS: